add print-certificate page
